[~] tests: consolidate svod-tensor/svod-model test additions
matmul.rs: factor the three AMD compile-only tensor-core tests onto shared
`matmul_kernel_ast` / `amd_optimizer` / `pinned_tc_config` / `render_amd`
helpers, and give the M=5 padded-WMMA regression a `Fragments` collector so
its A/B/C address coverage is asserted once and replayed at both the optimized
and linearized stages instead of being written out twice. Fold the shape-only
`dot`/`linear` tests into two `#[test_case]` tables, drop the three
`test_print_*_ir` println dumpers and the print block in the beam-search
regression, and merge the two all-ones matmuls into one table.

schedule.rs: add `lane_buffers`/`mstack_of`/`call_over`/`lane_inputs`/
`device_axis` helpers; merge the three MSTACK-expansion tests into one and the
four multi-device rejection tests into one table of error variants.

realize_internal.rs: `param_call_item` helper; the three
`resolve_compiled_kernel_buffer_indices` ordering tests become one table (they
asserted the same result three ways), the two `input_indices` tests become one
table, and the optimized-kernel-key test loses its repeated call literals.

memory_planner.rs: merge the arena/remap metrics tests and the pre-existing
`test_arena_reports_memory_savings` into one differential test.

Also: merge the empty float max/min identity tests (reduce.rs), compact the
transcendental table behind one closure (math.rs), assert the whole
`max_pool2d` output rather than element 0 (nn.rs), fold the unbound-variable
test into the existing `codegen_tests!` block (variable.rs), drop review-ID
prefixes (beam_worker.rs), and delete the two `test_nonzero_2d_debug_*`
eprintln scaffolds (indexing.rs) covered by `test_nonzero_2d`.

model: blocks_internal.rs shares an `fp16` builder and an `assert_rounds_once`
helper across the LayerNorm and Linear epilogue tests.

+++ tensor/src/test/unit/rangeify_kernel_count.rs
@@ -31,7 +31,7 @@ fn ln(x: &Tensor) -> Tensor {
 
 /// Layernorm with the per-channel affine the conformer conv module applies.
 fn ln_affine(x: &Tensor) -> Tensor {
-    let c = x.shape().expect("shape").last().and_then(|d| d.as_const()).expect("static channels") as usize;
+    let c = x.shape().expect("shape").last().and_then(|d| d.as_const()).expect("static channels");
     let (w, b) = (f32_in(&[c]), f32_in(&[c]));
     ln(x).try_mul(&w).expect("scale").try_add(&b).expect("shift")
 }
